safari-534.53-branch/LayoutTests/fast/url/trivial-segments-expected.txt

Test basic features of URL segmentation

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS segments('http://example.com/') is '["http:","example.com","0","/","",""]'
PASS successfullyParsed is true

TEST COMPLETE
